requirements.txt
pandas
plotly
yfinance
psutil
python-dateutil
kaleido
